pyproject.toml
[project]
name = "contracts-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.11"
dependencies = [
    "chromadb>=1.3.7",
    "huggingface-hub>=0.36.0",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "tiktoken>=0.12.0",
    "transformers>=4.57.3",
    "watchdog>=6.0.0",
]
